requirements.txt
tensorflow==2.15.0
gradio
numpy
Pillow